Created a gazebo plugin with the same functionality as cone_ground_truth.py.
Currently only the cone positions and their markers are being published, so not the midpoints.
Everything also still needs to use parameters.
I also still need to simulate data from the lidar and camera by adding noise to the ground truth data.
Finally I need to make this an option when launching the simulation.