Add noise object
The noise object takes in a state and returns a new state which has had
Gaussian noise applied to each field. In order to make this work nicely,
I have had to add in more fields to the state object (such as z) so that
I can publish them without manually adding noise.